Fixed build issue with tillstock (See CORE-1980)

See merge request securetrading-gl/retail-platform/retail-platform/cloud-office!192

+++ grails-app/domain/uk/co/wonderlane/wlpos/TillStock.groovy
@@ -24,7 +24,7 @@ class TillStock {
         model column: "model"
         retailerId column: "retailerId", sqlType: "tinyint"
         tillId column: "tillId", sqlType: "int"
-        storeId column: "storeId", sqlType: "int"
+        storeId column: "storeId", sqlType: "smallint"
         dateUpdated column: "dateUpdated"
     }
